requirements.txt
a2a-sdk==0.3.24
ag-ui-protocol==0.1.13
agent-framework==1.0.0rc3
agent-framework-a2a==1.0.0b260304
agent-framework-ag-ui==1.0.0b260304
agent-framework-anthropic==1.0.0b260304
agent-framework-azure-ai==1.0.0rc3
agent-framework-azure-ai-search==1.0.0b260304
agent-framework-azurefunctions==1.0.0b260304
agent-framework-bedrock==1.0.0b260304
agent-framework-chatkit==1.0.0b260304
agent-framework-claude==1.0.0b260304
agent-framework-copilotstudio==1.0.0b260304
agent-framework-core==1.0.0rc3
agent-framework-declarative==1.0.0b260304
agent-framework-devui==1.0.0b260304
agent-framework-durabletask==1.0.0b260304
agent-framework-foundry-local==1.0.0b260304
agent-framework-github-copilot==1.0.0b260304
agent-framework-lab==1.0.0b251024
agent-framework-mem0==1.0.0b260304
agent-framework-ollama==1.0.0b260304
agent-framework-orchestrations==1.0.0b260304
agent-framework-purview==1.0.0b260304
agent-framework-redis==1.0.0b260304
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
asttokens==3.0.1
async-timeout==5.0.1
asyncio==4.0.0
attrs==25.4.0
azure-ai-agents==1.2.0b5
azure-ai-inference==1.0.0b9
azure-ai-projects==2.0.0b4
azure-common==1.1.28
azure-core==1.38.2
azure-functions==1.24.0
azure-functions-durable==1.5.0
azure-identity==1.25.2
azure-search-documents==11.7.0b2
azure-storage-blob==12.28.0
backoff==2.2.1
boto3==1.42.63
botocore==1.42.63
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
claude-agent-sdk==0.1.48
click==8.3.1
clr_loader==0.2.10
colorama==0.4.6
comm==0.2.3
cryptography==46.0.5
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
docstring_parser==0.17.0
durabletask==1.3.0
durabletask.azuremanaged==1.3.0
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.135.1
foundry-local-sdk==0.5.1
frozenlist==1.8.0
furl==2.1.4
github-copilot-sdk==0.1.25
google-api-core==2.30.0
google-auth==2.49.0
googleapis-common-protos==1.73.0
greenlet==3.3.2
griffe==1.15.0
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
ipykernel==7.2.0
ipython==8.38.0
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
jsonpath-ng==1.8.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mcp==1.26.0
mem0ai==1.0.5
microsoft-agents-activity==0.8.0
microsoft-agents-copilotstudio-client==0.8.0
microsoft-agents-hosting-core==0.8.0
ml_dtypes==0.5.4
msal==1.35.1
msal-extensions==1.3.1
multidict==6.7.1
nest-asyncio==1.6.0
numpy==2.2.6
ollama==0.6.1
openai==2.26.0
openai-agents==0.10.5
openai-chatkit==1.6.3
opentelemetry-api==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opentelemetry-semantic-conventions-ai==0.4.15
orderedmultidict==1.0.2
packaging==26.0
parso==0.8.6
platformdirs==4.9.4
portalocker==3.2.0
posthog==7.9.7
powerfx==0.0.34
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
pure_eval==0.2.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
python-ulid==3.1.0
pythonnet==3.0.5
pytz==2026.1.post1
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
qdrant-client==1.17.0
redis==7.1.1
redisvl==0.15.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
sse-starlette==3.3.2
stack-data==0.6.3
starlette==0.52.1
tenacity==9.1.4
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
Werkzeug==3.1.6
yarl==1.23.0
zipp==3.23.0
